Refactor: infra/error 에 있는 공통 error에를 기준으로 컨트롤러 수정

+++ backend/src/main/java/com/grepp/smartwatcha/app/controller/api/details/InterestApiController.java
@@ -3,38 +3,41 @@ package com.grepp.smartwatcha.app.controller.api.details;
 
 import com.grepp.smartwatcha.app.model.auth.CustomUserDetails;
 import com.grepp.smartwatcha.app.model.details.service.jpaservice.InterestJpaService;
+import com.grepp.smartwatcha.infra.error.exceptions.CommonException;
 import com.grepp.smartwatcha.infra.jpa.enums.Status;
+import com.grepp.smartwatcha.infra.response.ResponseCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-@RequiredArgsConstructor
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/movies/{id}/interest")
 public class InterestApiController {
     private final InterestJpaService interestJpaService;
 
     @PostMapping
-    public ResponseEntity<Void> setInterestStatus(
+    public void setInterestStatus(
             @PathVariable("id") Long movieId,
             @RequestParam Status status,
             @AuthenticationPrincipal CustomUserDetails userDetails
             ){
         if(userDetails == null){
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            throw new CommonException(ResponseCode.BAD_REQUEST);
         }
         String userEmail = userDetails.getUsername();
         interestJpaService.saveOrUpdateInterest(userEmail, movieId, status);
-        return ResponseEntity.ok().build();
     }
     @DeleteMapping
-    public ResponseEntity<Void> deleteInterest(@PathVariable("id") Long movieId,
+    public void  deleteInterest(@PathVariable("id") Long movieId,
                                                @AuthenticationPrincipal CustomUserDetails userDetails) {
         Long userId = userDetails.getUser().getId();
+        if(userId == null){
+            throw new CommonException(ResponseCode.BAD_REQUEST);
+        }
         interestJpaService.deleteInterest(userId, movieId);
-        return ResponseEntity.noContent().build();
     }
 }
